refactor kanban context with drag and drop
